Check if container image tag already published
Problem: Repositories are immutable for security reasons.

To prevent blocking a deploy that has already published an image, this adds checks if the image tag (short gitsha) has already been published, and then skips re-attempting to build and publish it.

This is useful for a workflow branch that may already have been deployed.

Alternatively, could turn off immutable tags in staging.